# feat: optimize `lower` and `upper` functions