## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes https://github.com/apache/arrow-datafusion/issues/9970

## Rationale for this change

When converting case, there is no need to transform each value individually. Instead, you can treat the data in the StringArray as a single value and then perform the case conversion on it. Benefits include:

- Avoiding iteration.
- Directly allocating a contiguous block of memory, thereby avoiding the overhead of repeatedly allocating memory when creating a StringArray through an Iterator.

### Benchmark

#### Lower

```shell
Gnuplot not found, using plotters backend
lower full optimization: 1024
                        time:   [5.8336 µs 5.8401 µs 5.8481 µs]
                        change: [-77.800% -77.702% -77.608%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) low severe
  8 (8.00%) high mild
  6 (6.00%) high severe

lower maybe optimization: 1024
                        time:   [52.137 µs 52.206 µs 52.287 µs]
                        change: [-3.1792% -2.9748% -2.7567%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

lower partial optimization: 1024
                        time:   [28.266 µs 28.289 µs 28.316 µs]
                        change: [-47.687% -47.535% -47.391%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  5 (5.00%) low severe
  1 (1.00%) low mild
  6 (6.00%) high mild
  10 (10.00%) high severe

lower full optimization: 4096
                        time:   [23.473 µs 23.506 µs 23.549 µs]
                        change: [-77.596% -77.519% -77.434%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

lower maybe optimization: 4096
                        time:   [216.41 µs 216.65 µs 216.92 µs]
                        change: [-4.0400% -3.7733% -3.5015%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high severe

lower partial optimization: 4096
                        time:   [118.70 µs 118.87 µs 119.08 µs]
                        change: [-47.772% -47.681% -47.588%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  3 (3.00%) low severe
  4 (4.00%) low mild
  2 (2.00%) high mild
  9 (9.00%) high severe

lower full optimization: 8192
                        time:   [49.088 µs 49.232 µs 49.401 µs]
                        change: [-77.615% -77.497% -77.359%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high severe

lower maybe optimization: 8192
                        time:   [435.56 µs 435.96 µs 436.47 µs]
                        change: [-5.1301% -4.8973% -4.6614%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe

lower partial optimization: 8192
                        time:   [238.99 µs 239.86 µs 240.89 µs]
                        change: [-46.649% -46.348% -46.021%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

```

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->