@JasonLi-cn 2024-04-12
### New Test

```shell
Gnuplot not found, using plotters backend
lower_all_values_are_ascii: 1024
                        time:   [5.2583 µs 5.2623 µs 5.2670 µs]
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe

lower_the_first_value_is_nonascii: 1024
                        time:   [53.434 µs 53.474 µs 53.524 µs]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

lower_the_middle_value_is_nonascii: 1024
                        time:   [53.889 µs 53.975 µs 54.083 µs]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe

lower_all_values_are_ascii: 4096
                        time:   [20.936 µs 20.950 µs 20.965 µs]
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

lower_the_first_value_is_nonascii: 4096
                        time:   [222.68 µs 222.90 µs 223.17 µs]
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

lower_the_middle_value_is_nonascii: 4096
                        time:   [223.73 µs 223.98 µs 224.30 µs]
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe

lower_all_values_are_ascii: 8192
                        time:   [41.524 µs 41.796 µs 42.172 µs]
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

lower_the_first_value_is_nonascii: 8192
                        time:   [449.05 µs 449.71 µs 450.57 µs]
Found 13 outliers among 100 measurements (13.00%)
  4 (4.00%) high mild
  9 (9.00%) high severe

lower_the_middle_value_is_nonascii: 8192
                        time:   [451.06 µs 452.63 µs 454.66 µs]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
```